pyproject.toml
[build-system]
requires = ["hatchling", "uv-dynamic-versioning"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "uv-dynamic-versioning"

[tool.uv-dynamic-versioning]
vcs = "git"
style = "pep440"
bump = true

[project]
name = "omnimemory"
dynamic = ["version"]
description = "Advanced LLM-native memory orchestration with dual-agent synthesis, conflict-aware evolution, multi-tenant vector DB backends, and async SDK/API tooling."
readme = "README.md"
authors = [{ name = "Abiola Adeshina", email = "abiolaadedayo1993@gmail.com" }]
requires-python = ">=3.10"

keywords = ["memory", "vector-database", "llm", "embeddings", "chromadb", "qdrant", "mongodb", "pgvector", "ai", "multi-tenant"]
license = { text = "MIT" }
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "asyncpg>=0.29.0",
    "chromadb>=1.3.5",
    "fastapi>=0.115.0",
    "litellm>=1.79.3",
    "motor>=3.7.1",
    "prometheus-client>=0.23.1",
    "pydantic>=2.12.4",
    "pymongo>=4.15.4",
    "python-decouple>=3.8",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.15.1",
    "rapidfuzz>=3.14.3",
    "rich>=14.2.0",
    "tiktoken>=0.12.0",
    "typer>=0.12.0",
    "uvicorn[standard]>=0.32.0",
]

[project.scripts]
omnimemory = "omnimemory.cli.main:app"

[project.urls]
Repository = "https://github.com/omnirexflora-labs/omnimemory"
Issues = "https://github.com/omnirexflora-labs/omnimemory/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/omnimemory"]

[tool.hatch.build]
packages = ["src/omnimemory"]
include = [
    "LICENSE",
    "README.md"
]

[tool.pytest.ini_options]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "asyncio: marks tests that use asyncio",
    "OpenAIIntegration: marks tests that require an OpenAI API key",
]

[tool.uv.sources]
omnimemory = { workspace = true }

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.11.7",
    "hatch>=1.14.1",
    "twine>=6.1.0",
    "mypy>=1.11.0",
]

[tool.coverage.run]
source = ["src/omnimemory"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/conftest.py",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true
show_error_context = true
pretty = true
ignore_missing_imports = false
disable_error_code = ["import-untyped"]
namespace_packages = true
explicit_package_bases = true
mypy_path = "src"

[[tool.mypy.overrides]]
module = [
    "chromadb.*",
    "qdrant_client.*",
    "motor.*",
    "pymongo.*",
    "rapidfuzz.*",
    "tiktoken.*",
    "litellm.*",
    "prometheus_client.*",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "omnimemory.*"
warn_return_any = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
ignore_missing_imports = false
follow_imports = "normal"

[[tool.mypy.overrides]]
module = [
    "decouple.*",
    "asyncpg.*",
]
ignore_missing_imports = true

